Error and clean up

+++ server/app.py
@@ -1,9 +1,9 @@
 # Remote library imports
-from flask import request, make_response, session
+from flask import request, make_response, session, redirect
 from flask_restful import Resource
 import os
 from datetime import datetime
-from sqlalchemy.exc import IntegrityError
+from functools import wraps
 
 
 BASE_DIR = os.path.abspath(os.path.dirname(__file__))
@@ -58,6 +58,7 @@ class OrderItems(Resource):
             return make_response(serialized_order_items, 200)
         except Exception as e:
             return make_response({"error": str(e)}, 400)
+            return make_response({"error": str(e)}, 400)
 
     def post(self):
         try:
@@ -73,10 +74,16 @@ class OrderItems(Resource):
                 data = request.get_json()
                 total = 0
                 for order_item in data:
+                    quantity = int(order_item["quantity"])
                     product = db.session.get(Product, order_item["product_id"])
+                    if quantity > product.stock:
+                        return {
+                            "error": f"Requested quantity for {product.name} exceeds available stock"
+                        }, 400
+                    elif quantity == 0:
+                        return {"error": f"Quantity needs to be greater than 0."}, 400
                     new_order_item = OrderItem(**order_item, order_id=new_order.id)
                     db.session.add(new_order_item)
-                    db.session.commit()
                     product.stock -= int(order_item["quantity"])
                     db.session.commit()
                     total += new_order_item.price_at_order * new_order_item.quantity
@@ -117,6 +124,7 @@ class Login(Resource):
             user = User.query.filter_by(email=data.get("email")).first()
             if user and user.authenticate(data.get("password_hash")):
                 session["user_id"] = user.id
+                return make_response(user.to_dict(), 200)
                 return make_response(user.to_dict(), 200)
             else:
                 return make_response({"error": "Incorrect email or password"}, 401)
@@ -228,16 +236,19 @@ class Products(Resource):
             return make_response(serialized_products, 200)
         except Exception as e:
             return make_response({"error": str(e)}, 400)
+            return make_response({"error": str(e)}, 400)
 
     def post(self):
         try:
-            data = request.get_json()
-            new_product = Product(**data)
-            db.session.add(new_product)
-            db.session.commit()
+            if user_id := session.get("user_id"):
+                data = request.get_json()
+                new_product = Product(**data)
+                db.session.add(new_product)
+                db.session.commit()
             return make_response(new_product.to_dict(), 201)
         except Exception as e:
             db.session.rollback()
+            return make_response({"error": str(e)}, 400)
             return make_response({"error": str(e)}, 400)
 
 
@@ -247,42 +258,43 @@ class ProductById(Resource):
             product = db.session.get(Product, id)
 
             if not product:
-                return make_response({"error": str(e)}, 404)
-            else:
-                return make_response(product.to_dict(), 200)
+                return make_response({"error": "Product not found"}, 404)
+            return make_response(product.to_dict(), 200)
         except Exception as e:
+            return make_response({"error": str(e)}, 400)
             return make_response({"error": str(e)}, 400)
 
     def patch(self, id):
         try:
-            data = request.get_json()
-            product = db.session.get(Product, id)
-            if product:
-                for attr, value in data.items():
-                    if value:
-                        setattr(product, attr, value)
-                db.session.commit()
-                return make_response(product.to_dict(), 200)
-            return make_response({"error": "Product not found"}, 404)
+            if user_id := session.get("user_id"):
+                data = request.get_json()
+                product = db.session.get(Product, id)
+                if product:
+                    for attr, value in data.items():
+                        if value:
+                            setattr(product, attr, value)
+                    db.session.commit()
+                    return product.to_dict(), 200
+                return {"error": "Product not found"}, 400
+            return make_response({"error": "No logged in user"}, 401)
         except Exception as e:
             db.session.rollback()
             return make_response({"error": str(e)}, 422)
 
     def delete(self, id):
         try:
-            if product := db.session.get(Product, id):
-                db.session.delete(product)
-                db.session.commit()
-                return {}, 204
-            return make_response({"error": "Product not found"}, 404)
+            if user_id := session.get("user_id"):
+                if product := db.session.get(Product, id):
+                    db.session.delete(product)
+                    db.session.commit()
+                    return {}, 204
+                return make_response({"error": "Product not found"}, 400)
+            return make_response({"error": "No logged in user"}, 401)
         except Exception as e:
             db.session.rollback()
-            return make_response({"error": str(e)}, 422)
-
-        
+            return make_response({"error": str(e)}, 400)
 
 
-api.add_resource(Orders, "/orders")
 api.add_resource(OrderItems, "/order_items")
 api.add_resource(Signup, "/signup")
 api.add_resource(Login, "/login")
@@ -292,8 +304,6 @@ api.add_resource(UserById, "/user/<int:id>")
 api.add_resource(CheckSession, "/check-session")
 api.add_resource(Products, "/products")
 api.add_resource(ProductById, "/products/<int:id>")
-
-
 
 
 if __name__ == "__main__":
